requirements.txt
python-telegram-bot==20.7
reportlab==4.1.0
requests==2.31.0
python-dotenv==1.0.0
